refactor: convert float variables to use double

+++ src/main/java/com/BudgiePanic/rendering/io/CanvasWriter.java
@@ -60,7 +60,7 @@ public final class CanvasWriter {
      *     Builder, if the row has room for more colors.
      *     If the row has reached maximum length, the string builder is written to lines and a new SB is returned. 
      */
-    private static StringBuilder writeColor(StringBuilder builder, float color, List<String> lines) {
+    private static StringBuilder writeColor(StringBuilder builder, double color, List<String> lines) {
         final float min = 0f;
         final float max = 1f;
         final float scale = 255f;
