+++ src/main/java/com/BudgiePanic/rendering/util/light/PointLight.java
@@ -16,9 +16,9 @@ import com.BudgiePanic.rendering.util.Tuple;
 public record PointLight(Tuple position, Color color) implements Light {
 
     @Override
-    public float intensityAt(Tuple point, World world, float time) {
+    public double intensityAt(Tuple point, World world, double time) {
         final boolean inShadow = world.isOccluded(point, position, World.shadowCasters, time);
-        return inShadow ? 0f : 1f;
+        return inShadow ? 0.0 : 1.0;
     }
 
     @Override
